streamlit_requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
PyGithub>=2.1.1
langchain-groq>=0.1.0
langchain>=0.3.24
langchain-google-genai>=2.1.3
langchain-groq>=0.1.0
langchain_community>=0.0.10
tavily-python>=0.3.0
streamlit>=1.28.0
langchain_community>=0.0.10
langchain-openai>=0.1.0
fastembed>=0.2.0 